requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
boto3==1.35.24
requests==2.32.3
pillow==10.4.0
python-dotenv==1.0.1
python-multipart==0.0.9
